Moved filters to separate component
